+++ src/components/EditListingFeaturesPanel/EditListingFeaturesPanel.js
@@ -10,7 +10,7 @@ import { ListingLink } from '../../components';
 
 import css from './EditListingFeaturesPanel.css';
 
-const FEATURES_NAME = 'amenities';
+const FEATURES_NAME = 'filters';
 
 const EditListingFeaturesPanel = props => {
   const {
@@ -39,8 +39,8 @@ const EditListingFeaturesPanel = props => {
     <FormattedMessage id="EditListingFeaturesPanel.createListingTitle" />
   );
 
-  const amenities = publicData && publicData.amenities;
-  const initialValues = { amenities };
+  const filters = publicData && publicData.filters;
+  const initialValues = { filters };
 
   return (
     <div className={classes}>
@@ -50,10 +50,10 @@ const EditListingFeaturesPanel = props => {
         name={FEATURES_NAME}
         initialValues={initialValues}
         onSubmit={values => {
-          const { amenities = [] } = values;
+          const { filters = [] } = values;
 
           const updatedValues = {
-            publicData: { amenities },
+            publicData: { filters },
           };
           onSubmit(updatedValues);
         }}