+++ src/components/EditListingWizard/EditListingWizard.js
@@ -87,7 +87,7 @@ const tabCompleted = (tab, listing) => {
     case DESCRIPTION:
       return !!(description && title);
     case FEATURES:
-      return !!(publicData && publicData.amenities);
+      return !!(publicData && publicData.filters);
     case POLICY:
       return !!(publicData && typeof publicData.rules !== 'undefined');
     case LOCATION: