+++ src/components/FieldCheckboxGroup/FieldCheckboxGroup.example.js
@@ -8,7 +8,7 @@ import { requiredFieldArrayCheckbox } from '../../util/validators';
 const formName = 'Styleguide.FieldCheckboxGroup';
 const formNameRequired = 'Styleguide.FieldCheckboxGroupRequired';
 
-const label = <h3>Amenities</h3>;
+const label = <h3>Filters</h3>;
 
 const commonProps = {
   label: label,
@@ -54,14 +54,14 @@ const commonProps = {
 };
 
 const optionalProps = {
-  name: 'amenities-optional',
-  id: 'amenities-optional',
+  name: 'filters-optional',
+  id: 'filters-optional',
   ...commonProps,
 };
 
 const requiredProps = {
-  name: 'amenities-required',
-  id: `${formNameRequired}.amenities-required`,
+  name: 'filters-required',
+  id: `${formNameRequired}.filters-required`,
   ...commonProps,
   validate: requiredFieldArrayCheckbox('this is required'),
 };