+++ src/forms/EditListingFeaturesForm/EditListingFeaturesForm.js
@@ -56,7 +56,7 @@ const EditListingFeaturesFormComponent = props => (
             className={css.features}
             id={name}
             name={name}
-            options={config.custom.amenities}
+            options={config.custom.filters}
           />
 
           <Button