+++ src/marketplace-custom-config.js
@@ -56,3 +56,11 @@ export const priceFilterConfig = {
 export const dateRangeFilterConfig = {
   active: true,
 };
+
+// Activate keyword filter on search page
+
+// NOTE: If you are ordering search results by distance the keyword search can't be used at the same time.
+// You can turn off ordering by distance in config.js file
+export const keywordFilterConfig = {
+  active: true,
+};